# refactor(Barretenberg): Static analyzer optimization