There are fixes in the code which accelerated constructor of static analyzer.

During testing, we noticed that graph creation was slow in some cases(SHA256 NIST vector five & join split circuit). Using perf utility we found that some parts of the code were ineffective. 

SHA256 NIST vector five test case showed that the most time-consuming part was transformation lookup tables columns in std::set, which was unnecessary, because we didn't need to order elements, we just wanted to get the amount of unique elements in the columns, so it would be better to use std::unordered_set in this case. No this test works ~90 secs(93 secs and 87.30, so approximately 90 seconds).

Join Split Circuit Test showed that the most time-consuming part was checking that variable wasn't constant and the problem was getting constant_variable_indices from Ultra Circuit Builder, because there wasn't reference for this object. Now test with all removing parts works 97 seconds whereas it works 2 minutes and 26 seconds before.